Remove `Vec` from `DescriptorSetElementWrite` struct.
Each `DescriptorSetElementWrite` now represents a single index with a binding. If a binding has multiple elements, there will be one `DescriptorSetElementWrite` for each element. `DescriptorSetElementKey` now contains the `array_index` field representing which element the write is targeting. For cases where that doesn't matter, see `DescriptorSetBindingKey` containing just the `dst_binding` field.

The `MaterialInstanceSlotAssignment` now contains the `array_index` so that it can be used to create a `DescriptorSetElementKey`.