BUGFIX: Pass shadow-specific per-view descriptor set.
Commit 563707 wasn't passing the per-view descriptor sets to the shadow pass and no shadows were drawn as a result.

It didn't make sense to share the same per-view uniform as the opaque pass because the shadow pass doesn't need any of the lighting data, so this commit defines new shadow uniforms in `mesh_shadow_map.glsl`. As there's no usage of most of the definitions in `mesh.glsl`, I removed the import of that entirely. This greatly reduced the size of the generated code for `mesh_shadow_map.vert`. In the demo's `Mesh` feature, `per_view_descriptor_set` is no longer an `Option` because it's required for both the opaque & the shadow draw calls. The shadow map view data is collected in `shadow_per_view_descriptor_set_layouts`, populated by the `"create per view dsecriptor sets"` scope, and ultimately assigned to each node at the `add_render_node` call-site. The `is_shadow_pass` argument was removed from `add_render_node` because it wasn't used anymore.