Add depth pre-pass.
This change removes the shadow pass from the `mesh.material` and instead defines a new `depth.material`. This assumes that all shadows can use a predefined material for the shadow map. The new `depth.material` is passed through the `mesh/extract` job into `mesh/prepare` where it is used for both shadows and the depth pre-pass.