Refactor text feature.
If a feature exposes structs for use outside of the jobs, those should go into a `public` folder and be re-exported inside of the `mod`.
If a feature needs to define common functionality for itself that doesn't fit into any of the existing jobs, put that into an `internal` folder and don't export it outside of the feature.