Add `RenderFeatureMask` and `RenderFeatureMaskBuilder`.
Views can now filter out features using a `RenderFeatureMask` in the same way as they can filter on phases with the `RenderPhaseMask`. When building the frame packet, we'll filter out visible nodes that don't match the view's feature mask. The `sprite_scene` shows an example of disabling the skybox by not registering the main view with that feature.

This commit also fixes a crash in the nodes API example demo feature.