Update demo.
This commit replaces any code associated with the `DynamicVisibilityNodeSet` or `StaticVisibilityNodeSet` with the new `VisibilityRegion` defined by `rafx-framework`.

- It also adds the calculation of the `VisibleBounds` struct to the `gltf` importer and the calculation of layer bounds to the `ltdk` importer.
- The `PositionComponent` is replaced with a more general `TransformComponent` component and the fields that would otherwise have been duplicated are removed from `SpriteRenderNode` or `MeshRenderNode`. Each type of `Component` for supporting a light now defines the `ViewFrustumArc` required for the `RenderViews` used for that light's shadow maps.
- The `extract` job for the `Mesh` feature only queries `Entities` in the `Legion` world that have an entry in the frame packet.
- The `extract` job for the `Sprite` feature only queries `Entities` in the `Legion` world that have an entry in the frame packet.
- The `TileLayerResource` now registers itself with the `VisibilityRegion` in `set_project`.

Each of the demos has been updated to use the `VisibilityRegion`.

## rafxmark

This demo is a benchmark on a worst-case scenario where all entities are visible to the orthogonal main view and require per-frame position updates. The position updates could be batched with a single asynchronous command, but this demo instead submits separate asynchronous commands for each entity inside of a `Legion` system. This demo also shows how to register a new object with the `VisibilityRegion` from inside of a `Legion` system.

## ManySprites

This demo is a benchmark on culling a large number of entities. There are 102400 sprites in the scene but nearly all of them are culled from the orthogonal main view during each frame.

## SpriteScene

This demo culls the parts of the `LDtk` scene that are not visible to the orthogonal main view.

## ShadowsScene

This demo demonstrates frustum culling on multiple `ViewFrustums` representing the main view and the 3 shadow-casters in the scene -- 1 overhead directional light, 1 spot light, and 6 faces of the cube map for the 1 point light rotating the scene.